+++ SwiftStylable/Classes/Style/StyleSets/PlaceholderStyle.swift
@@ -24,9 +24,9 @@ public class PlaceholderStyle : StyleSetBase {
     //
     // -----------------------------------------------------------------------------------------------------------------------
     
-    internal init(name:String, parent:PlaceholderStyle? = nil, data:[String:Any], colorCollection:ColorCollection) {
+    internal init(name:String, parent:PlaceholderStyle? = nil, data:[String:Any], colorCollection:ColorCollection, dimensionCollection:DimensionCollection) {
         self._parent = parent
-        super.init(name: name, parent: parent, colorCollection: colorCollection)
+        super.init(name: name, parent: parent, colorCollection: colorCollection, dimensionCollection: dimensionCollection)
         self.applyData(data)
     }