+++ SwiftStylable/Classes/Style/Stylers/ButtonTextStylable.swift
@@ -5,8 +5,7 @@
 //  Created by Marcel Bloemendaal on 17/04/2018.
 //
 
-import Foundation
-
+import UIKit
 
 protocol ButtonTextStylable : AnyObject {
 	var titleLabel:UILabel? { get }