+++ SwiftStylable/Classes/Style/Stylers/ForegroundStyleable.swift
@@ -5,7 +5,7 @@
 //  Created by Marcel Bloemendaal on 17/04/2018.
 //
 
-import Foundation
+import UIKit
 
 protocol ForegroundStylable : class {
 	var foregroundColor: UIColor? { get set }